==> experiments/experiment_2/notebooks/0_exp2_concrete_rating.ipynb <==
import pandas as pd
import os
# Current work
print(os.getcwd())
# Change directory 2 level up
os.chdir("..")
os.chdir("..")
os.chdir("..")

print(os.getcwd())
# Look for the file in the directory
print(os.listdir())

file_path = 'experiments/experiment_2/data/13428_2013_403_MOESM1_ESM.xlsx'

try:
    concreteness_data = pd.read_excel(file_path, sheet_name=0)
    
    preview = concreteness_data.head()
except Exception as e:
    preview = str(e)

preview


## looking and processing data

from experiments.experiment_2.data import words

MWB1=words.MATERIAL_WELL_BEING['Concrete']
MWB1 = [word if word != "books" else "book" for word in MWB1] # fix at the end replacing book with books
MWB2=words.MATERIAL_WELL_BEING['Abstract']

H1=words.HEALTH['Concrete']
H1 = [word if word != "cigarettes" else "cigarette" for word in H1] # fix at the end replacing cigarette with cigarettes
H1 = [word if word != "fruits" else "fruit" for word in H1] # fix at the end replacing fruit with fruits

H2=words.HEALTH['Abstract']

P1=words.PRODUCTIVITY['Concrete']
P2=words.PRODUCTIVITY['Abstract']
print(P1)
P1= [word if word != "time-management" else "timekeeping" for word in P1] # fix at the end replacing timekeeping with time-management
print(P1)
I1=words.INTIMACY['Concrete']
I1= [word if word != "bickering" else "bicker" for word in I1] # fix at the end replacing bickering with squabbling
I2=words.INTIMACY['Abstract']
I1= [word if word != "sharing" else "share" for word in I1] # fix at the end replacing bickering with squabbling
print(I1)

ALL_LIST = MWB1 + MWB2 + H1 + H2 + P1 + P2 + I1 + I2
#imprimir len de all list
print(len(ALL_LIST))
ALL_LIST_CONCRETE = words.MATERIAL_WELL_BEING['Concrete'] + words.HEALTH['Concrete'] + words.PRODUCTIVITY['Concrete'] + words.INTIMACY['Concrete']
ALL_LIST_ABSTRACT = words.MATERIAL_WELL_BEING['Abstract'] + words.HEALTH['Abstract'] + words.PRODUCTIVITY['Abstract'] + words.INTIMACY['Abstract']

### ALL LIST

filtered_data = concreteness_data[concreteness_data['Word'].isin(ALL_LIST)]
filtered_data = filtered_data.sort_values('Word')

missing_words = set(ALL_LIST) - set(filtered_data['Word'])

filtered_data, missing_words

## Median-split

median_concreteness = filtered_data['Conc.M'].median()

filtered_data["ConcretenessCategory"] = filtered_data["Conc.M"].apply(lambda x: "Concrete" if x >= median_concreteness else "Abstract")

filtered_data['Word'] = filtered_data['Word'].apply(lambda x: 'books' if x == 'book' else x)
filtered_data['Word'] = filtered_data['Word'].apply(lambda x: 'cigarettes' if x == 'cigarette' else x)
filtered_data['Word'] = filtered_data['Word'].apply(lambda x: 'fruits' if x == 'fruit' else x)
filtered_data['Word'] = filtered_data['Word'].apply(lambda x: 'time-management' if x == 'timekeeping' else x)
filtered_data['Word'] = filtered_data['Word'].apply(lambda x: 'bickering' if x == 'bicker' else x)
filtered_data['Word'] = filtered_data['Word'].apply(lambda x: 'sharing' if x == 'share' else x)

concrete_words = filtered_data[filtered_data['ConcretenessCategory'] == 'Concrete']['Word'].head(80).tolist()
abstract_words = filtered_data[filtered_data['ConcretenessCategory'] == 'Abstract']['Word'].head(80).tolist()

dict_r = {
    'concrete': concrete_words,
    'abstract': abstract_words
}


median_concreteness, filtered_data[['Word', 'Conc.M', 'ConcretenessCategory']].sort_values(by='Conc.M')

dict_r
#imprimir el len de cada categoria
print(len(concrete_words))
print(len(abstract_words))

## Comparison

### MWB

all_set_concrete = set(ALL_LIST_CONCRETE)
all_set_abstract = set(ALL_LIST_ABSTRACT)

result_all_set_concrete = set(dict_r['concrete'])
result_all_set_abstract = set(dict_r['abstract'])

diferencias_concrete = all_set_concrete.difference(result_all_set_concrete)
diferencias_abstract = all_set_abstract.difference(result_all_set_abstract)

print("Words in the original list that are not in the result list (concrete):", diferencias_concrete)
print("Words in the original list that are not in the result list (abstract):", diferencias_abstract)

## Save data

import json

with open('experiments/experiment_2/data/concreteness.json', 'w') as f:
    json.dump(dict_r, f, indent=4)
