==> experiments/experiment_2/notebooks/1_change_category.ipynb <==
import os
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import numpy as np
import gensim.downloader as api
import numpy as np
from sklearn.decomposition import PCA
import matplotlib.pyplot as plt 

# Current work
print(os.getcwd())
# Change directory 2 level up
os.chdir("..")
os.chdir("..")
os.chdir("..")

print(os.getcwd())
# Look for the file in the directory
print(os.listdir())

model = "llama"
prompt_technique = "react"
temperature = 0.3
data = pd.read_csv(f"experiments/experiment_2/results/{model}/results_{prompt_technique}_{model}_{temperature}.csv")


concreteness = pd.read_json('experiments/experiment_2/data/concreteness.json')

def determine_category(word):
    if word in concreteness['concrete'].values:
        return 'Concrete'
    elif word in concreteness['abstract'].values:
        return 'Abstract'
    else:
        return 'Unknown'

# apply low case to the 'Surveyed Word' column
data['Surveyed Word'] = data['Surveyed Word'].str.lower()

# if in data in surveyede word column a word is Time management, then change it to time-management
data['Surveyed Word'] = data['Surveyed Word'].str.replace(' ', '-')
data['Surveyed Word'] = data['Surveyed Word'].replace('management', 'time-management')
data["Related 1"] = data["Related 1"].str.lower()
data["Related 2"] = data["Related 2"].str.lower()
data["Related 3"] = data["Related 3"].str.lower()
data["Related 4"] = data["Related 4"].str.lower()
data["Related 5"] = data["Related 5"].str.lower()
data["Related 6"] = data["Related 6"].str.lower()
data["Related 7"] = data["Related 7"].str.lower()
data["Related 8"] = data["Related 8"].str.lower()
data["Related 9"] = data["Related 9"].str.lower()
data["Related 10"] = data["Related 10"].str.lower()

data['Category'] = data['Surveyed Word'].apply(determine_category)

data.to_csv(f"experiments/experiment_2/results/{model}/results_{prompt_technique}_{model}_{temperature}_cleaned.csv", index=False)

concreteness['concrete']

#check len of concrete and abstract from the variable concreteness
print(len(concreteness['concrete']))
print(len(concreteness['abstract']))

#check how many unknown are in surveyed word column
print(len(data[data['Category'] == 'Unknown']))
