==> experiments/experiment_2/notebooks/6_exp2_pca_explained.ipynb <==
import pandas as pd
from sklearn.decomposition import PCA
import numpy as np
import matplotlib.pyplot as plt
import os

# Current work
print(os.getcwd())
# Change directory 2 level up
os.chdir("..")
os.chdir("..")
os.chdir("..")
print(os.getcwd())
# Look for the file in the directory
print(os.listdir())

model = "llama"
prompt_technique = "chain"
temperature = 0.7

data = pd.read_csv(f'experiments/experiment_2/results_with_distances/{model}/results_with_distances_{prompt_technique}_{model}_{temperature}_distances.csv')
data[data['Iteration'] == 1]

distances_columns = [col for col in data.columns if 'Sequential Distance' in col]
numeric_data = data[['Induced Pathology', 'Surveyed Word'] + distances_columns]
grouped_data = numeric_data.groupby(['Induced Pathology', 'Surveyed Word']).mean().reset_index()

pca_results = []
induced_pathologies = grouped_data['Induced Pathology'].unique()
variance_by_pathology_iteration = {}
n_components = 10

for pathology in induced_pathologies:
    variance_by_pathology_iteration[pathology] = []
    for iteration in data['Iteration'].unique():
        iteration_data = data[(data['Induced Pathology'] == pathology) & (data['Iteration'] == iteration)]
        sequential_distances_data = iteration_data.filter(like='Sequential Distance').dropna()
        if len(sequential_distances_data) >= n_components:
            pca = PCA(n_components=n_components)
            pca.fit(sequential_distances_data)
            varianza_acumulada = np.cumsum(pca.explained_variance_ratio_) * 100
            variance_by_pathology_iteration[pathology].append(varianza_acumulada)

for pathology in variance_by_pathology_iteration:
    variance_by_pathology_iteration[pathology] = np.array(variance_by_pathology_iteration[pathology])

marcadores = {'Depression': 'o', 'Trait Anxiety': 's', 'Eating Disorder': '^', 'Alcohol Addiction': 'D',
              'Impulsivity': 'v', 'Schizophrenia': '<', 'Obsessive Compulsive Disorder': '>', 'Apathy': 'P',
              'Social Anxiety': '*', 'No Pathology': 'X'}

fig, ax = plt.subplots(figsize=(14, 8))
colores = plt.cm.tab10(np.linspace(0, 1, len(induced_pathologies)))

for i, pathology in enumerate(variance_by_pathology_iteration):
    mean_variance = np.mean(variance_by_pathology_iteration[pathology], axis=0)
    std_variance = np.std(variance_by_pathology_iteration[pathology], axis=0)
    components = np.arange(1, n_components + 1)
    ax.plot(components, mean_variance, label=pathology, color=colores[i], marker=marcadores[pathology])
    ax.fill_between(components, mean_variance - std_variance, mean_variance + std_variance, color=colores[i], alpha=0.1)

ax.axhline(y=80, color='r', linestyle='--', label='80% Variance Explained')

ax.set_xlabel('Dimensions', fontsize=22)
ax.set_ylabel('Variance explained', fontsize=22)
ax.set_title('Semantic Dimensionality', fontsize=25)
ax.legend(loc='best', fontsize=18)

ax.set_xticks(np.arange(1, 11))

ax.tick_params(axis='both', which='major', labelsize=18)

plt.tight_layout()
plt.savefig(f'experiments/experiment_2/figs/{model}/{temperature}/exp2_semantic_dimensionality_{prompt_technique}_{model}_{temperature}.png')
plt.show()
